Changed the way app id checkbox lists are built.  Get unicode working
Got unicode working correctly.  Just had to make sure the database knew about
it, as well as json and a couple other things.

App id checkboxes are now just a flat list of type 'AppId'.  This change also
has the beginnings of the 'Server Options' checkboxes but I wanted to get the
fixed unicode in.